pyproject.toml
[tool.poetry]
name = "sl-transformer"
version = "0.1.0"
description = ""
authors = ["Cleison Amorim <cca5@cin.ufpe.br>"]

[tool.poetry.dependencies]
python = "^3.8"
# commons-python = { git = "https://github.com/amorim-cleison/commons-python.git", branch = "develop" }
commons-python = { path = "../commons-python/", develop = false }
torchtext= "^0.6.0"
spacy = "^2.3.2"
seaborn = "^0.11.0"
torch = "1.7.1"
sklearn = "^0.0"
numpy = "1.21.0"
pandas = "^1.3.2"
skorch = "^0.10.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
yapf = "^0.30.0"
pytest = "^6.1.1"
rope = "^0.18.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

